+++ Base/PLRenderer/src/Texture/Creator/TextureCreatorNormalizationCubeMap.cpp
@@ -122,7 +122,7 @@ Texture *TextureCreatorNormalizationCubeMap::Create(TextureManager &cTextureMana
 	// Create the normalization cube map data
 	Image cImage;
 	for (uint8 i=0; i<6; i++) {
-		ImageBuffer *pImageBuffer = cImage.CreatePart(ImagePartCubeSidePosX + i)->CreateMipmap();
+		ImageBuffer *pImageBuffer = cImage.CreatePart(static_cast<uint32>(ImagePartCubeSidePosX) + i)->CreateMipmap();
 		pImageBuffer->CreateImage(DataByte, ColorRGB, Vector3i(Size, Size, 1));
 
 		uint8 *d = pImageBuffer->GetData();
